pyproject.toml
[project]
name = "graphcut-test"
version = "0.1.0"
description = "A Python project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "open3d>=0.19.0",
    "pymaxflow>=1.3.2",
    "scikit-learn>=1.8.0",
]
